+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 
 - Set "company_id" params as not required for address/update_address resource
 - Ignore "/" and "/doc" in resource/get_resources resource
+- Restructure the Entity page on oxe-web-admin
 
 ### Fixed